+++ local/newassignment_remover/CHANGELOG.md
@@ -3,3 +3,10 @@
 ## 3.9.0-r0.1 (2022-01-20)
 
 - First beta release
+
+## 3.9.0-r0.2 (2O22-03-10)
+
+- Fixed grade types not transfering
+- Fixed incorrect submission type mapping
+- Changed "Attempts reopened" assign option to map to "Manually" instead of "Never"
+- Extracted conversion scripts to be able to be used as automatic tool.